Qt-KDE Wrapper (http://iidea.pl/~js/qkw/)  code (KDElibs for win32) merged with current kdecore/
important notes:

-fixx11h.h    -unused outside X11
-kaccelmanager_private.h  -KAccelManagerAlgorithm static members
                           changed to static
-kapplication.cpp   -commented out code related to
                      ICE, X11, SM
-kaudioplayer.cpp    -could work if only KNotifyClient::userEvent() could
                      worked
-kconfigbase.cpp   -- QDir::isRelativePath(dir) used; "/etc/kderc" mapped to
                      $WINDIR/kderc
-kcrash.cpp        -untested, ifdefed out some code
-kextsock.cpp    --not ported yet
-kglobalaccel_emb.h    --unused for win32
-kglobalaccel_x11.cpp  --unused for win32
-kglobalaccel_x11.h    --unused for win32
-kidna.cpp    -ifdefed only, depends on knetwork
-kipc.cpp      -ifdefed onlym, because used X11 for IPC
-kkeynative_x11.cpp  -ported, maybe rename it to kkeynative.cpp?
-kkeyserver_x11.cpp  -ported, maybe rename it to kkeyserver.cpp?
-kkeyserver_x11.h    -ported, maybe rename it to kkeyserver.h?
-kmanagerselection.cpp   --not ported yet
-kmountpoint.cpp   -ifdefed one place
-kprocctrl.cpp -untested, ifdefed out some code
-kprocess.cpp  -untested, ifdefed out some code
-kpty.cpp          -ifdefed out
-kregexp.cpp  --not included (deprecated), use QRegExp instead
-kregexp.h    --not included (deprecated), use QRegExp instead
-kregpriv.h   --not included (deprecated), use QRegExp instead
-krootprop.cpp   -unused
-ksock.cpp      --not ported yet
-ksocks.cpp     --not ported yet
-kstandarddirs.cpp  ---kfsstnd_defaultbindir(), kfsstnd_defaultprefix()
                       added
-kstartupinfo.cpp    --X11 only, ifdefed
-kstartupinfo.h      --X11 only
-kurl.cpp     -mostly working, may need fixing for special cases with drive
               letters
-kvmallocator.cpp  --not ported, we can use win32 wrapper for mmap here
-kwin.cpp    --Unix only
-kwinmodule.cpp  --Unix only
-kwinmodule.h    --Unix only
-kxerrorhandler.cpp --X11 only
-kxerrorhandler.h  --X11 only
-kxmessages.cpp   --X11 only
-kxmessages.h     --X11 only
-netsupp.cpp      --added empty impl. in netsupp_win32.cpp
-netwm.cpp   --X11 only
-netwm.h     --X11 only
-netwm_def.h --X11 only
-netwm_p.h   --X11 only

svn path=/trunk/kdelibs/; revision=344656

+++ kdecore/kcharsets.cpp
@@ -469,7 +469,7 @@ QString KCharsets::resolveEntities( const QString &input )
 
         const uint ampersandPos = ampersand - text.unicode();
 
-        text[ ampersandPos ] = entityValue;
+        text[ (int)ampersandPos ] = entityValue;
         text.remove( ampersandPos + 1, entityLength + 1 );
         p = text.unicode() + ampersandPos;
         end = text.unicode() + text.length();
